split tables
