+++ iservices/iotrxserver.go
@@ -1,0 +1,3 @@
+package iservices
+
+const IOTrxServiceName = "iotrx_svc"
